set up unit tests and fix serve-dev runner
set up unit tests for all packages. Fix serve-dev.js to not try to
live-serve packages that don't have a live-server conf. This allows
running serve-dev as a all-in-one command for building and testing
with file watching, even for geolocator package. Otherwise you
would have to run build -w and test -w separately.